+++ tools/Azure.Mcp.Tools.AzureManagedLustre/tests/Azure.Mcp.Tools.AzureManagedLustre.LiveTests/AzureManagedLustreCommandTests.cs
@@ -29,7 +29,7 @@ namespace Azure.Mcp.Tools.AzureManagedLustre.LiveTests
         public async Task Should_calculate_required_subnet_size()
         {
             var result = await CallToolAsync(
-                "azmcp_azuremanagedlustre_filesystem_required-subnet-size",
+                "azmcp_azuremanagedlustre_filesystem_subnetsize_ask",
                 new()
                 {
                     { "subscription", Settings.SubscriptionId },
@@ -94,5 +94,43 @@ namespace Azure.Mcp.Tools.AzureManagedLustre.LiveTests
             }
         }
 
+
+        [Fact]
+        public async Task Should_check_subnet_size_and_succeed()
+        {
+            var result = await CallToolAsync(
+                "azmcp_azuremanagedlustre_filesystem_subnetsize_validate",
+                new()
+                {
+                    { "subscription", Settings.SubscriptionId },
+                    { "sku", "AMLFS-Durable-Premium-40" },
+                    { "size", 480 },
+                    { "location", Environment.GetEnvironmentVariable("LOCATION") },
+                    { "subnet-id", Environment.GetEnvironmentVariable("AMLFS_SUBNET_ID") }
+                });
+
+            var valid = result.AssertProperty("valid");
+            Assert.Equal(JsonValueKind.True, valid.ValueKind);
+            Assert.True(valid.GetBoolean());
+        }
+
+        [Fact]
+        public async Task Should_check_subnet_size_and_fail()
+        {
+            var result = await CallToolAsync(
+                "azmcp_azuremanagedlustre_filesystem_subnetsize_validate",
+                new()
+                {
+                    { "subscription", Settings.SubscriptionId },
+                    { "sku", "AMLFS-Durable-Premium-40" },
+                    { "size", 1008 },
+                    { "location", Environment.GetEnvironmentVariable("LOCATION") },
+                    { "subnet-id", Environment.GetEnvironmentVariable("AMLFS_SUBNET_SMALL_ID") }
+                });
+
+            var valid = result.AssertProperty("valid");
+            Assert.Equal(JsonValueKind.False, valid.ValueKind);
+            Assert.False(valid.GetBoolean());
+        }
     }
 }